+++ example/windows/flutter/generated_plugins.cmake
@@ -6,11 +6,10 @@ list(APPEND FLUTTER_PLUGIN_LIST
   connectivity_plus
   media_kit_libs_windows_video
   media_kit_video
-  share_plus
-  url_launcher_windows
 )
 
 list(APPEND FLUTTER_FFI_PLUGIN_LIST
+  jni
 )
 
 set(PLUGIN_BUNDLED_LIBRARIES)